Enough parsing to need memory mgmt
This adds some basic memory mgmt to deal with the recursive data
structures. Maybe basic is too kind a term; I'm stuffing the data onto
the heap (with a new for every struct :sad:) and storing the pointers on
a couple of vectors which are (pinky promise) cleaned up by the
destructor.

I'm thinking of making this more elaborite (not better, mind you) by
using another variant so we can have just one heap object because I love
std::visit so much apparently.

Note that this is a working commit - it's missing the rest of the
chapter so it won't compile.