More parse rules
This is where I remembered that I was _that guy_ and used
a pointer to LiteralVal too. The rule of three dictates
that it's time to refactor the heap into variants.

If you only had this codebase to go off of, you'd think
that 99% of problems are solved with variants.